feat:refactor code

+++ src/pages/Order/Order.jsx
@@ -2,6 +2,11 @@ import React, { useEffect, useState } from "react";
 import "./order.scss";
 import { Link } from "react-router-dom";
 import { publicRequest } from "../../requestMethod";
+
+import {AiOutlineSearch} from 'react-icons/ai'
+import {MdDone} from 'react-icons/md'
+import {HiXMark} from 'react-icons/hi2'
+
 import axios from "axios";
 const Order = () => {
   const [enableModelCreate, setEnableModelCreate] = useState(false);
@@ -40,30 +45,39 @@ const Order = () => {
 
   const handleConfirm = async (id) => {
     try {
-      await axios.get(
-        `http://localhost:8080/api/v1/travel/order/update/${id}`,
-        {
-          headers: {
-            Authorization: `Bearer ${
-              JSON.parse(localStorage.getItem("user"))?.accessToken || null
-            }`,
-          },
-        }
-      );
-      getOrder();
+
+      if(window.confirm('Xác nhận đơn đặt lịch ?') === true){   
+          await axios.get(
+            `http://localhost:8080/api/v1/travel/order/update/${id}`,
+            {
+              headers: {
+                Authorization: `Bearer ${
+                  JSON.parse(localStorage.getItem("user"))?.accessToken || null
+                }`,
+              },
+            }
+          );
+
+          getOrder();
+      }
+
     } catch (error) {}
   };
 
   const handleDelete = async (id) => {
     try {
-      await axios.delete(`http://localhost:8080/api/v1/travel/order/${id}`, {
-        headers: {
-          Authorization: `Bearer ${
-            JSON.parse(localStorage.getItem("user"))?.accessToken || null
-          }`,
-        },
-      });
-      getOrder();
+
+      if(window.confirm('Bạn có chắc muốn xóa người đơn đặt lịch này không ?') === true){  
+          await axios.delete(`http://localhost:8080/api/v1/travel/order/${id}`, {
+            headers: {
+              Authorization: `Bearer ${
+                JSON.parse(localStorage.getItem("user"))?.accessToken || null
+              }`,
+            },
+          });
+          getOrder();
+      }
+
     } catch (error) {}
   };
 
@@ -79,10 +93,9 @@ const Order = () => {
         style={{ display: enableModelCreate === true && "block" }}
       ></div>
 
-      <h1>Đơn hàng</h1>
+      <h1 style={{height:40}}></h1>
       <table id="customers">
         <tr>
-          <th>ID người dùng</th>
           <th>Tên người dùng</th>
           <th>Email người dùng</th>
           <th>Địa chỉ người dùng</th>
@@ -96,7 +109,7 @@ const Order = () => {
           order.map((item) => {
             return (
               <tr>
-                <td>{item.userId}</td>
+                
                 <td>{item.customerName}</td>
                 <td>{item.customerEmail}</td>
                 <td>{item.customerAddress}</td>
@@ -109,27 +122,19 @@ const Order = () => {
                     justifyContent: "space-between",
                   }}
                 >
-                  <button className="btn-of-order">
+                 
                     <Link
                       className="link"
                       to={`/travel/order/details/${item.id}`}
                     >
-                      Chi tiết
+                      <AiOutlineSearch  size={20}/>
                     </Link>
-                  </button>
+                  
 
-                  <button
-                    className="btn-of-order"
-                    onClick={() => handleConfirm(item.id)}
-                  >
-                    Xác nhận
-                  </button>
-                  <button
-                    className="btn-of-order"
-                    onClick={() => handleDelete(item.id)}
-                  >
-                    Hủy
-                  </button>
+                
+                    <MdDone style={{cursor:'pointer',color:'green'}} size={20} onClick={() => handleConfirm(item.id)}/>
+                    <HiXMark style={{cursor:'pointer',color:'red'}} size={20} onClick={() => handleDelete(item.id)} />
+                
                 </td>
               </tr>
             );
